reorganize tensorflow files
