--- alibi_detect/models/__init__.py
@@ -1,12 +1,0 @@
-from .autoencoder import AE, AEGMM, VAE, VAEGMM
-from .resnet import resnet
-from .pixelcnn import PixelCNN
-
-__all__ = [
-    "AE",
-    "AEGMM",
-    "VAE",
-    "VAEGMM",
-    "resnet",
-    "PixelCNN"
-]
