+++ alibi_detect/models/tensorflow/__init__.py
@@ -1,0 +1,12 @@
+from .autoencoder import AE, AEGMM, VAE, VAEGMM
+from .pixelcnn import PixelCNN
+from .resnet import resnet
+
+__all__ = [
+    "AE",
+    "AEGMM",
+    "VAE",
+    "VAEGMM",
+    "resnet",
+    "PixelCNN"
+]
