+++ alibi_detect/models/tensorflow/losses.py
@@ -2,7 +2,7 @@ import tensorflow as tf
 from tensorflow.keras.layers import Flatten
 from tensorflow.keras.losses import kld, categorical_crossentropy
 import tensorflow_probability as tfp
-from alibi_detect.models.gmm import gmm_params, gmm_energy
+from alibi_detect.models.tensorflow.gmm import gmm_params, gmm_energy
 
 
 def elbo(y_true: tf.Tensor,
