+++ alibi_detect/models/tests/test_losses.py
@@ -1,6 +1,6 @@
 import numpy as np
 import tensorflow as tf
-from alibi_detect.models.losses import elbo, loss_adv_ae, loss_aegmm, loss_vaegmm, loss_distillation
+from alibi_detect.models.tensorflow.losses import elbo, loss_adv_ae, loss_aegmm, loss_vaegmm, loss_distillation
 
 N, K, D, F = 10, 5, 1, 3
 x = np.random.rand(N, F).astype(np.float32)
