+++ alibi_detect/od/ae.py
@@ -2,8 +2,8 @@ import logging
 import numpy as np
 import tensorflow as tf
 from typing import Dict, Tuple
-from alibi_detect.models.autoencoder import AE
-from alibi_detect.models.trainer import trainer
+from alibi_detect.models.tensorflow.autoencoder import AE
+from alibi_detect.models.tensorflow.trainer import trainer
 from alibi_detect.base import BaseDetector, FitMixin, ThresholdMixin, outlier_prediction_dict
 from alibi_detect.utils.prediction import predict_batch
