+++ alibi_detect/od/vaegmm.py
@@ -3,10 +3,10 @@ import numpy as np
 import tensorflow as tf
 import tensorflow_probability as tfp
 from typing import Callable, Dict, Tuple
-from alibi_detect.models.autoencoder import VAEGMM, eucl_cosim_features
-from alibi_detect.models.gmm import gmm_energy, gmm_params
-from alibi_detect.models.losses import loss_vaegmm
-from alibi_detect.models.trainer import trainer
+from alibi_detect.models.tensorflow.autoencoder import VAEGMM, eucl_cosim_features
+from alibi_detect.models.tensorflow.gmm import gmm_energy, gmm_params
+from alibi_detect.models.tensorflow.losses import loss_vaegmm
+from alibi_detect.models.tensorflow.trainer import trainer
 from alibi_detect.base import BaseDetector, FitMixin, ThresholdMixin, outlier_prediction_dict
 from alibi_detect.utils.prediction import predict_batch
